- Added more features

+++ src/main/java/itsix/CreditProject/models/IMoney.java
@@ -2,6 +2,6 @@ package itsix.CreditProject.models;
 
 public interface IMoney {
 
-	Integer getValue();
+	Double getValue();
 
 }
